Move generic methods from organisations to client
