+++ app/services/schools/dfe_sign_in_api/client.rb
@@ -8,6 +8,28 @@ module Schools
           ENV['DFE_SIGNIN_API_ENDPOINT']
         ].map(&:presence).all?
       end
+
+    private
+
+      def response
+        resp = Faraday.new.get(endpoint) do |req|
+          req.headers['Authorization'] = "bearer #{token}"
+          req.headers['Content-Type']  = 'application/json'
+        end
+
+        JSON.parse(resp.body)
+      end
+
+      def token
+        JWT.encode(payload, ENV.fetch('DFE_SIGNIN_API_SECRET'))
+      end
+
+      def payload
+        {
+          iss: ENV.fetch('DFE_SIGNIN_API_CLIENT'),
+          aud: 'signin.education.gov.uk'
+        }
+      end
     end
   end
 end
