+++ app/services/schools/dfe_sign_in_api/organisations.rb
@@ -25,31 +25,11 @@ module Schools
         @organisations ||= response
       end
 
-      def response
-        resp = Faraday.new.get(endpoint) do |req|
-          req.headers['Authorization'] = "bearer #{token}"
-          req.headers['Content-Type']  = 'application/json'
-        end
-
-        JSON.parse(resp.body)
-      end
-
       def endpoint
         URI::HTTPS.build(
           host: Rails.configuration.x.dfe_sign_in_api_host,
           path: ['/users', user_uuid, 'organisations'].join('/')
         )
-      end
-
-      def token
-        JWT.encode(payload, ENV.fetch('DFE_SIGNIN_API_SECRET'))
-      end
-
-      def payload
-        {
-          iss: ENV.fetch('DFE_SIGNIN_API_CLIENT'),
-          aud: 'signin.education.gov.uk'
-        }
       end
     end
   end
